+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Let current_request respond to .html as well as .json
+- Make it work also when Rack::Attack.cache.store is set to an instance of ActiveSupport::Cache::RedisCacheStore (from rails)
 
 ## 0.1.1 (2019-04-02)